// shared/constants/types/rpc-stellar-gen.d.ts
export type Boolean = boolean
export type Bytes = Buffer
export type Double = number
export type Int = number
export type Int64 = number
export type Long = number
export type String = string
// ...
export type AccountID = String
// ...
export type BuildPaymentResLocal = {
  readonly readyToReview: Boolean
  readonly from: AccountID
  readonly toErrMsg: String
  readonly amountErrMsg: String
  readonly secretNoteErrMsg: String
  readonly publicMemoErrMsg: String
  readonly worthDescription: String
  readonly worthInfo: String
  readonly worthAmount: String
  readonly worthCurrency: String
  readonly displayAmountXLM: String
  readonly displayAmountFiat: String
  readonly sendingIntentionXLM: Boolean
  readonly amountAvailable: String
  readonly banners?: Array<SendBannerLocal> | null
}
export type BuildRequestResLocal = {
  readonly readyToRequest: Boolean
  readonly toErrMsg: String
  readonly amountErrMsg: String
  readonly secretNoteErrMsg: String
  readonly worthDescription: String
  readonly worthInfo: String
  readonly displayAmountXLM: String
  readonly displayAmountFiat: String
  readonly sendingIntentionXLM: Boolean
  readonly banners?: Array<SendBannerLocal> | null
}
// ...
export type SendBannerLocal = {
  readonly level: String
  readonly message: String
  readonly proofsChanged: Boolean
}